+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
+- Remove task logs on `pueue remove`. [#187](https://github.com/Nukesor/pueue/issues/187)
+- Improve Windows support by [oiatz](https://github.com/oiatz). [#114](https://github.com/Nukesor/pueue/issues/114)
+
 ## [0.12.1] - 12-03-2021
 
 ### Fixed